refactor(exception): move global handler
