+++ src/main/java/com/classroomscheduler/exception/GlobalExceptionHandler.java
@@ -1,0 +1,36 @@
+package com.classroomscheduler.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import java.util.Map;
+
+@RestControllerAdvice
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(RecursoNaoEncontradoException.class)
+    public ResponseEntity<Map<String, String>> handleNotFound(RecursoNaoEncontradoException exception) {
+        return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                .body(Map.of("erro", exception.getMessage()));
+    }
+
+    @ExceptionHandler(RegraDeNegocioException.class)
+    public ResponseEntity<Map<String, String>> handleBadRequest(RegraDeNegocioException exception) {
+        return ResponseEntity.badRequest()
+                .body(Map.of("erro", exception.getMessage()));
+    }
+
+    @ExceptionHandler(NaoAutorizadoException.class)
+    public ResponseEntity<Map<String, String>> handleUnauthorized(NaoAutorizadoException exception) {
+        return ResponseEntity.status(HttpStatus.UNAUTHORIZED)
+                .body(Map.of("erro", exception.getMessage()));
+    }
+
+    @ExceptionHandler(AcessoNegadoException.class)
+    public ResponseEntity<Map<String, String>> handleForbidden(AcessoNegadoException exception) {
+        return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                .body(Map.of("erro", exception.getMessage()));
+    }
+}
